+++ src/sketches/templates/template.rs
@@ -18,38 +18,21 @@ pub const SKETCH_CONFIG: SketchConfig = SketchConfig {
 #[derive(SketchComponents)]
 pub struct Template {
     hub: ControlHub<Timing>,
-    radius: f32,
     hue: f32,
 }
 
 pub fn init(_app: &App, ctx: &LatticeContext) -> Template {
-    let controls = ControlHubBuilder::new()
+    let hub = ControlHubBuilder::new()
         .timing(Timing::new(ctx.bpm()))
-        .slider("radius", 100.0, (10.0, 500.0), 1.0, None)
+        .slider("radius", 100.0, (10.0, 300.0), 1.0, None)
         .build();
 
-    Template {
-        hub: controls,
-        radius: 0.0,
-        hue: 0.0,
-    }
+    Template { hub, hue: 0.0 }
 }
 
 impl Sketch for Template {
-    fn update(&mut self, _app: &App, _update: Update, ctx: &LatticeContext) {
-        let radius_max = self.hub.float("radius");
-
-        self.radius = self.hub.animation.automate(
-            &[
-                Breakpoint::ramp(0.0, 10.0, Easing::Linear),
-                Breakpoint::ramp(1.0, ctx.window_rect().hw(), Easing::Linear),
-                Breakpoint::ramp(2.0, 10.0, Easing::Linear),
-                Breakpoint::ramp(3.0, radius_max, Easing::Linear),
-                Breakpoint::end(4.0, 10.0),
-            ],
-            Mode::Loop,
-        );
-
+    fn update(&mut self, _app: &App, _update: Update, _ctx: &LatticeContext) {
+        self.hub.update();
         self.hue = self.hub.animation.tri(12.0)
     }
 
@@ -64,7 +47,7 @@ impl Sketch for Template {
 
         draw.ellipse()
             .color(hsl(self.hue, 0.5, 0.5))
-            .radius(self.radius)
+            .radius(self.hub.get("radius"))
             .x_y(0.0, 0.0);
 
         draw.to_frame(app, &frame).unwrap();